appium-tests/requirements.txt
Appium-Python-Client>=2.0.0
pytest>=7.0.0
selenium>=4.0.0
